Create Sum of two large number
To find the sum of the two large number. it is not possible if we use two numbers as integers, as only 4bits are assigned to integers. Hence we take the input as string then perform the operating by converting single character to integer and the sum and store in the array. And finally converting the final array to string.